+++ src/main/java/com/baemin/repositories/ClubNumRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.baemin.domain.entity.ClubNum;
 
-public interface ClubNumRepository extends JpaRepository<ClubNum, Long> {
+public interface ClubNumRepository extends JpaRepository<ClubNum, Integer> {
 	List<ClubNum> findByStartYear(int year);
 	ClubNum findBySemester(int semester);
 	ClubNum findByStartYearAndSemester(int startYear, int semester);
