+++ src/main/java/com/baemin/repositories/JoinClubRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.baemin.domain.entity.JoinClub;
 
-public interface JoinClubRepository extends JpaRepository<JoinClub, Long> {
+public interface JoinClubRepository extends JpaRepository<JoinClub, Integer> {
 
 }
